requirements.txt
# Main dependencies
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
# For production
gunicorn

# Database and ORM
SQLAlchemy==2.0.21
# greenlet is installed separately in install_dependencies.sh

# Web requests and scraping
requests>=2.32.2
beautifulsoup4==4.12.2
playwright==1.38.0

# AI features
google-generativeai==0.3.2
# nearai==0.1.1

# Utilities
python-dotenv==1.0.0
werkzeug==2.3.6
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
docx2txt
spacy==3.7.2
psycopg2